requirements.txt
matplotlib
networkx==2.3
PyYAML
torch==2.0.1
tqdm
